feat: customLogger support info and error log
